客户端增加消息回调

+++ jforgame-socket/src/main/java/jforgame/socket/client/CallBackService.java
@@ -27,7 +27,7 @@ public class CallBackService {
         mapper.put(correlationId, future);
     }
 
-    public RequestResponseFuture remove(long correlationId) {
+    public RequestResponseFuture remove(int correlationId) {
         return mapper.remove(correlationId);
     }
